Migrated all endpoints to phoenix
